fix: resolve bug caused by refactor
