+++ chaekit-spring/src/main/resources/application.yml
@@ -26,7 +26,6 @@ spring:
     locations: classpath:db/migration
     baseline-on-migrate: false
     baseline-version: 1
-    repair-on-migrate: true
   mail:
     host: smtp.gmail.com
     port: 587
